requirements.txt
# requirements.txt for zeroscratches + BPBTL (Python 3.10.19)
# Core deep learning
torch==2.0.1
torchvision==0.15.2

# BPBTL dependencies
dlib==19.24.2
scikit-image==0.25.2
easydict==1.13
PyYAML==6.0.3
dominate>=2.3.1
dill==0.3.8
tensorboardX==2.6.4
scipy==1.11.4
opencv-python==4.8.1.78
einops==0.8.1
PySimpleGUI==5.0.8.3
matplotlib==3.7.3

# zeroscratches dependencies
numpy==1.24.3
Pillow==10.1.0
huggingface-hub==0.20.3
zeroscratches>=0.1.0

# Shared utilities
tqdm==4.66.1
requests==2.31.0